making quorum report as an integer

+++ member_management/views.py
@@ -28,5 +28,5 @@ def member_list(request):
     data['full_members'] = Person.objects.full_members().order_by('last_name')
     data['starving_hackers'] = Person.objects.starving_hackers().order_by('last_name')
     data['member_count'] = Person.objects.members().count()
-    data['quorum_count'] = max(ceil(float(data['full_members'].count() / 3)), 1)
+    data['quorum_count'] = max(int(ceil(float(data['full_members'].count() / 3))), 1)
     return render(request, 'member_management/member_list.html', data)
